Fixed issue with prepared statements.
The postgres driver's internal list of previously prepared statements
was static, which was a problem becuase other instances of the driver
created their own DB connections which did not have these statements
prepared. List of prepared statements changed to local member.

The Kwerry object had a bug where it would create multiple connections
for each instance, even for duplicate DSNs.  This caused the postgres
driver to try to use previously prepared statements on a connection
that did not have them prepared.